+++ R/mp_get_routes.R
@@ -4,14 +4,21 @@
 #' @export
 #' @examples
 #' library(xml2)
-#' doc = as_xml_document(response_directions)
+#' doc = as_xml_document(response_directions_driving)
+#' r = mp_get_routes(doc)
+#' plot(r)
+#' doc = as_xml_document(response_directions_transit)
 #' r = mp_get_routes(doc)
 #' plot(r)
 #' \dontrun{
+#' # Duration in traffic (only with API key)
+#' key = readLines("~/key") # API key
 #' doc = mp_directions(
 #'   origin = c(34.81127, 31.89277),
 #'   destination = c(34.781107, 32.085003),
-#'   alternatives = TRUE
+#'   departure_time = Sys.time(),
+#'   alternatives = TRUE,
+#'   key = key
 #' )
 #' r = mp_get_routes(doc)
 #' plot(r)
@@ -66,18 +73,22 @@ mp_get_routes = function(doc)  {
       xml_find_all(sprintf("/DirectionsResponse/route[%s]/leg/duration_in_traffic/value", i)) %>%
       xml_text %>%
       as.numeric
+    if(length(duration_in_traffic_s) == 0)
+      duration_in_traffic_s = NA
 
     duration_in_traffic_text =
       doc %>%
       xml_find_all(sprintf("/DirectionsResponse/route[%s]/leg/duration_in_traffic/text", i)) %>%
       xml_text
+    if(length(duration_in_traffic_text) == 0)
+      duration_in_traffic_text = NA
 
     rt = lapply(route, decode_line)
-    rt = lapply(rt, st_linestring)
-    rt = lapply(rt, st_sfc, crs = 4326)
+    rt = lapply(rt, sf::st_linestring)
+    rt = lapply(rt, sf::st_sfc, crs = 4326)
     rt = do.call(c, rt)
 
-  routes[[i]] = st_sf(
+  routes[[i]] = sf::st_sf(
     alternative_id = i,
     summary = summary,
     distance_m = distance_m,